+++ previous-versions/assess/src/sdk/sdk.ts
@@ -56,9 +56,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "1.0";
-    sdkVersion = "0.35.0";
-    genVersion = "2.159.2";
-    userAgent = "speakeasy-sdk/typescript 0.35.0 2.159.2 1.0 @codat/assess";
+    sdkVersion = "0.36.0";
+    genVersion = "2.169.0";
+    userAgent = "speakeasy-sdk/typescript 0.36.0 2.169.0 1.0 @codat/assess";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);