+++ previous-versions/assess/RELEASES.md
@@ -617,3 +617,13 @@ Based on:
 - [typescript v0.35.0] previous-versions/assess
 ### Releases
 - [NPM v0.35.0] https://www.npmjs.com/package/@codat/assess/v/0.35.0 - previous-versions/assess
+
+## 2023-10-23 01:23:25
+### Changes
+Based on:
+- OpenAPI Doc 1.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Assess.yaml
+- Speakeasy CLI 1.104.0 (2.169.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.36.0] previous-versions/assess
+### Releases
+- [NPM v0.36.0] https://www.npmjs.com/package/@codat/assess/v/0.36.0 - previous-versions/assess